:new: #2784【开放平台】增加批量代云开发之环境管理相关接口

+++ weixin-java-miniapp/src/main/java/cn/binarywang/wx/miniapp/api/WxMaShopOrderService.java
@@ -19,7 +19,7 @@ public interface WxMaShopOrderService {
 
   WxMaShopBaseResponse orderPay(WxMaShopOrderPayRequest request) throws WxErrorException;
 
-  WxMaShopGetOrderResponse getOrder(Integer orderId, String outOrderId, String openid)
+  WxMaShopGetOrderResponse getOrder(Long orderId, String outOrderId, String openid)
     throws WxErrorException;
